feat(init): Compute the current academic week number
Hardcoded is the Sunday of Week 0 date. In the future we should have an interface for the Admin to

set the "current" reference date. Or for the most overengineered and zero-maintenance, call out to a

port and scrape http://blink.ucsd.edu/instructors/resources/academic/calendars/YYYY.html for dates.

Looks like the first days of instruction are always Thursday Week 0, Monday Week 1, and Monday Week

1 for Fall, Winter, and Spring, respectively.